+++ 06.bowling_object/shot.rb
@@ -5,9 +5,18 @@ class Shot
     @mark = mark
   end
 
-  def score
+  def change_mark_to_score
     return 10 if mark == 'X'
 
     mark.to_i
+  end
+
+  def convert_mark_to_score
+    input_marks = ARGV[0]
+    add_0_after_markX = input_marks.gsub(/X/, 'X,0')
+    modified_scores = add_0_after_markX.gsub(/(?<=.{19})X0/, 'X')
+    modified_scores.split(",").map do |mark|
+      Shot.new(mark).change_mark_to_score
+    end
   end
 end
